Refactored admin-relevant components and state
